Rename skelanimdirector file to SkeletalTransformation and move from renderer into utils subfolder
