+++ plugins/guacamole-skelanim/include/gua/utils/SkeletalTransformation.hpp
